refactor: add use of slugify to specific folder, update test setup
