+++ packages/api-headless-cms/src/utils/toSlug.ts
@@ -1,0 +1,10 @@
+import slugify from "slugify";
+
+const toSlug = text =>
+    slugify(text, {
+        replacement: "-",
+        lower: true,
+        remove: /[*#\?<>_\{\}\[\]+~.()'"!:;@]/g
+    });
+
+export default toSlug;
